Separate signal interruption test
Signal interruption only runs on linux systems
